% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm-functions.R
\name{add_caption_to_plot}
\alias{add_caption_to_plot}
\title{Add Caption to Plot}
\usage{
add_caption_to_plot(p, caption_addition, sep = "\\n")
}
\arguments{
\item{p}{A ggplot object}

\item{caption_addition}{The caption to add}

\item{sep}{The separator between the original caption and the addition}
}
\value{
The caption for the ggplot
}
\description{
Add Caption to Plot
}
