% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{builtenvir}
\alias{builtenvir}
\alias{builtenvir-package}
\title{dlm: Distributed lag models in R using lme4}
\description{
dlm: Distributed lag models in R using lme4
}
\references{
Baek J, et al. (2016) Epidemiology 27(1):116-24.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/26414942}{PubMed})
}
