% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/methods-dlMod.R
\docType{methods}
\name{changePoint}
\alias{changePoint}
\alias{changePoint,dlMod-method}
\title{Lag Coefficient Change Points}
\usage{
changePoint(object, ...)

\S4method{changePoint}{dlMod}(object, ...)
}
\value{
An integer vector
}
\section{Methods (by class)}{
\itemize{
\item \code{dlMod}: S4 Method for \code{\link{dlMod}} objects
}}

