% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpi}
\alias{cpi}
\title{CPI}
\format{
A data frame with columns "year" and "cpi".
}
\source{
FRED
}
\usage{
data(cpi)
}
\description{
This dataset contains CPI data from FRED
}
\keyword{datasets}
