% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bases.R
\name{cr}
\alias{cr}
\title{Cubic Radial Basis}
\usage{
cr(x, Z, ...)
}
\arguments{
\item{x}{a vector of values to construct the basis from. Missing values
are not allowed.}

\item{Z}{a covariate matrix (or object that can be coerced to a \code{matrix})
to take the basis transformation of.
\code{length(x)} should be the same as \code{ncol(Z)}.
Missing values are not allowed.}

\item{...}{arguments to be passed to \code{\link{lag.basis}}}
}
\value{
An S4 object of class \code{\link{SmoothBasis}}.
}
\description{
Construct a natural cubic radial basis function for a given
distance set vector and apply as a linear transformation of a
covariate matrix.
}
\details{
At the time of writing, \code{cr} is little more than a convenient
  wrapper to \code{\link{lag.basis}} intended to simplify the task of
  specifying lag terms in a model \code{formula}. The longer term goal
  is for \code{cr} (and potentially other basis functions) to replace
  \code{lag.basis} entirely.
}
