% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm-functions.R
\name{distributed_lags_model}
\alias{distributed_lags_model}
\title{Distributed Lags Model}
\usage{
distributed_lags_model(
  data,
  exposure_data,
  from_rt,
  to_rt,
  outcome,
  exposure,
  unit,
  time,
  covariates = NULL,
  addl_fes = NULL,
  ref_period = -1,
  weights = NULL,
  dd = F,
  n = 2,
  dict = NULL
)
}
\arguments{
\item{data}{A data frame containing the unit, time, outcome, covariates, and any additional fixed effects}

\item{exposure_data}{A data frame containing the unit, time, and exposure variables}

\item{from_rt}{The starting lag period.}

\item{to_rt}{The ending lag period.}

\item{outcome}{The outcome variable.}

\item{exposure}{The exposure variable.}

\item{unit}{The unit identifier.}

\item{time}{The time variable.}

\item{covariates}{Vector of covariates for the model.}

\item{addl_fes}{Vector of additional fixed effects for the model.}

\item{ref_period}{Reference period (default -1)}

\item{weights}{Weights to be included in the regression}
}
\value{
A list containing model results, coefficients, and plots.
}
\description{
This is the distributed lags model / continuous event study.
}
