% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm-package.R
\docType{package}
\name{dlm-package}
\alias{dlm}
\alias{dlm-package}
\title{dlm: Distributed Lag Models for Event Studies}
\description{
Implements distributed lag models equivalent to event studies with binned endpoints, following Schmidheiny and Siegloch (2023, Journal of Applied Econometrics). Provides tools for estimation, visualization, and comparison with standard two-way fixed effects specifications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tlcaputi/dlm}
  \item Report bugs at \url{https://github.com/tlcaputi/dlm/issues}
}

}
\author{
\strong{Maintainer}: Theodore Caputi \email{tcaputi@gmail.com}

}
\keyword{internal}
