% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{dlMod}
\alias{dlMod}
\title{Distributed lag models}
\section{Slots}{

\describe{
\item{\code{resp}}{An \code{\link[lme4]{lmResp}} object to store a
mixed-model response variable}

\item{\code{bases}}{A list of \code{\link{LagBasis}} objects corresponding
to the unique set of lag bases used to fit the model}

\item{\code{index}}{A (named) integer vector providing the index of the
basis set in \code{bases} corresponding to each lag term in the
model}
}}

