% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/methods-dlMod.R
\docType{methods}
\name{estimands}
\alias{estimands}
\alias{Sigma}
\alias{vcoef}
\alias{vcoef0}
\alias{Sigma,dlMod-method}
\alias{vcoef0,dlMod-method}
\alias{vcoef,dlMod-method}
\alias{coef.dlMod}
\alias{confint.dlMod}
\title{Extract DLM parameter estimates}
\usage{
coef(object, ...)

confint(object, ...)

Sigma(object, ...)

vcoef(object, ...)

vcoef0(object, ...)

\S4method{Sigma}{dlMod}(object, scaled = TRUE, ...)

\S4method{vcoef0}{dlMod}(object, scaled = TRUE, ...)

\S4method{vcoef}{dlMod}(object, scaled = TRUE, ...)

\method{coef}{dlMod}(object, scaled = TRUE, ...)

\method{confint}{dlMod}(object, parm, level = 0.95, scaled = TRUE,
  coef = TRUE, ...)
}
\arguments{
\item{object}{a fitted \code{\link{dlMod}} object}

\item{...}{additional arguments}

\item{scaled}{if \code{TRUE} (the default), any lag parameters are scaled for
more natural interpretation (see Details)}

\item{parm}{an integer or character index to subset parameters}

\item{level}{the desired confidence level}

\item{coef}{if \code{coef = TRUE} (the default), \code{confint.dlMod} will
include an extra column (with \code{colname} \code{"coef"}) for the
regression coefficients themselves}
}
\value{
All of these functions return \code{numeric} data
}
\description{
Extract regression coefficients, variances, etc. from
fitted \code{\link{dlMod}} objects
}
\details{
Other typical methods like
\code{residuals}, and \code{sigma}, etc. are handled via inheritance
from \pkg{lme4} classes.
If the argument \code{scaled = TRUE}, parameter estimates are scaled
by the areas between radii and summed so that they can be interpreted
as the estimate up to a given radius (e.g. see the \pkg{dlmBE}
\link[=dlmBE-package]{package documentation}).

\describe{
  \item{\code{coef.dlMod}}{follows the format of
    \code{lme4::\link[lme4]{coef.merMod}} to return the sums of fixed
    and random effects for each level and grouping factor}
  \item{\code{confint.dlMod}}{returns confidence intervals for regression
    coefficients following \code{stats::\link[stats]{confint}}}
  \item{\code{Sigma}}{returns the regression coefficient covariance matrix.
    Row and column indices are in the same order as \code{vcoef} (see below)}
  \item{\code{vcoef}}{returns vectorized coefficients from the fitted
    model. Fixed effects come before distributed lag coefficients, which
    come before other random effects coefficients. For example, if
    \eqn{\beta} is a vector of fixed effects; \eqn{\theta_1} and
    \eqn{\theta_2} are vectors of (separately penalized) DL coefficients;
    and \eqn{b_1}, \eqn{b_2}, \eqn{\ldots}{...} are additional random
    effects vectors for groups \eqn{1, 2, \ldots}{1, 2, ...}, then
    \code{vcoef} will return the (named) vector
    \eqn{(\beta^T, \theta_1^T, \theta_2^T, b_1^T, b_2^T, \ldots)^T}{(\beta', \theta_1', \theta_2', b_1', b_2', ...)'}.
    \code{vcoef0} returns the same coefficient vector
    but without names (and is slightly faster).}
}
}
