% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm-functions.R
\name{generate_data}
\alias{generate_data}
\title{Generate Data}
\usage{
generate_data(seed = 1234, n_groups = 26^2, n_times = 20, treat_prob = 0.4)
}
\arguments{
\item{seed}{Seed for the random data}

\item{n_groups}{Number of groups}

\item{n_times}{Number of time periods}

\item{treat_prob}{Probability of treatment}
}
\value{
A dataset
}
\description{
Produces a dataset perfect for a standard TWFE model
}
