% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm-functions.R
\name{get_caption_from_plot}
\alias{get_caption_from_plot}
\title{Get Caption from Plot}
\usage{
get_caption_from_plot(p)
}
\arguments{
\item{p}{A ggplot object}
}
\value{
The caption for the ggplot
}
\description{
Get Caption from Plot
}
