% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.dlm.R
\name{interpret.dlm}
\alias{interpret.dlm}
\title{Interpret a DLM Formula}
\usage{
interpret.dlm(formula, data, .names.func = function(n) paste("pseudoGroups",
  n, sep = ""))
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted.
See \code{\link{dlm}} for details.}

\item{data}{a model-frame containing the data for each term in the model.
Should already be appropriately subset, etc.}
}
\value{
A \code{list} object...
}
\description{
Given an appropriate DL model formula and data-frame, prepares
the data to be fit by interpreting the formula and extracting
available smooth lag terms.
}
\details{
Uses R's \code{\link[stats]{model.matrix}} mechanisms to build
and parse "random effects" components of distributed lag terms.
}
