% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iplot_data.R
\name{iplot_data}
\alias{iplot_data}
\title{Internal function for grabbing and preparing iplot data}
\usage{
iplot_data(
  object,
  .ci_level = 0.95,
  .keep = NULL,
  .drop = NULL,
  .dict = fixest::getFixest_dict(),
  .aggr_es = c("none", "post", "pre", "both")
)
}
\arguments{
\item{object}{A model object of class `fixest` or `fixest_multi`, where
the `i()` operator has been used to construct an interaction, or set of
interactions.}

\item{.ci_level}{A number between 0 and 1 indicating the desired confidence
level, Defaults to 0.95.}

\item{.keep}{Character vector used to subset the coefficients of interest.
Passed down to `fixest::iplot(..., keep = .keep)` and should take the form of
an acceptable regular expression.}

\item{.drop}{Character vector used to subset the coefficients of interest
(complement of `.keep`). Passed down to `fixest::iplot(..., drop = .drop)`
and should take the form of an acceptable regular expression.}

\item{.dict}{A dictionary (i.e. named character vector or a logical scalar).
Used for changing coefficient names. Defaults to the values in
`getFixest_dict()`. See the `?fixest::iplot` documentation for more
information.}

\item{.aggr_es}{A character string indicating whether the aggregated mean
post- (and/or pre-) treatment effect should be added as a column to the
returned data frame. Passed to `aggr_es(..., aggregation = "mean")` and
should be one of "none" (the default), "post", "pre", or "both".}
}
\value{
A data frame consisting of estimate values, confidence intervals,
relative x-axis positions, and other aesthetic information needed to draw
a ggplot2 object.
}
\description{
Grabs the underlying data used to construct `fixest::iplot`,
with some added functionality and tweaks for the `ggiplot` equivalents.
}
\details{
This function is a wrapper around
`fixest::iplot(..., only.params = TRUE)`, but with various checks and tweaks
to better facilitate plotting with `ggplot2` and handling of complex object
types (e.g. lists of fixest_multi models)
}
\examples{
library(fixest)

est_did = feols(y ~ x1 + i(period, treat, 5) | id+period,
                data = base_did)
iplot(est_did, only.params = TRUE) # The "base" version
iplot_data(est_did)                # The wrapper provided by this package

# Illustrative fixest_multi case, where the sample has been split by odd and
# even ID numbers.
est_split = feols(y ~ x1 + i(period, treat, 5) | id+period,
                  data = base_did, split = ~id\%\%2)
iplot(est_split, only.params = TRUE) # The "base" version
iplot_data(est_split)                # The wrapper provided by this package

}
\seealso{
[fixest::iplot()], [aggr_es()].
}
