% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{LagBasis}
\alias{LagBasis}
\title{Create and manipulate lag basis functions}
\description{
S4 class object to store and query components of computed
lag basis functions. User interface for creating this class can be found in
\code{\link{basis}}.
}
\details{
See \code{\link{basis}} for details of the decomposition
}
\section{Slots}{

\describe{
\item{\code{x}}{radii that define ring-shaped areas around participant locations}

\item{\code{x.center}}{the value \code{x} was centered to prior to computing the basis}

\item{\code{x.scale}}{the value \code{x} was scaled by prior to computing the basis}

\item{\code{C0}}{\eqn{C_0} part of basis matrix}

\item{\code{K1}}{\eqn{K_1} part of basis matrix}

\item{\code{dist.fun}}{the function applied to the elements of \code{x} to compute the basis}
}}

\references{
Rupert D, Wand MP, & Carroll RJ (2003) Semiparametric
Regression. New York: Cambridge University Press.
}
