% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/methods-dlMod.R
\docType{methods}
\name{lagIndex}
\alias{lagIndex}
\alias{lagIndex,dlMod-method}
\title{Extract list of indices of lag terms}
\usage{
lagIndex(object, ...)

\S4method{lagIndex}{dlMod}(object, .fixed = TRUE, ...)
}
\section{Methods (by class)}{
\itemize{
\item \code{dlMod}: S4 Method for \code{\link{dlMod}} objects
}}

