% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm.R
\name{makeDlMod}
\alias{makeDlMod}
\alias{makeDlMod.merMod}
\title{Convert fitted models to 'dlMod' objects}
\usage{
\method{makeDlMod}{merMod}(object, parsed, call, ...)
}
\arguments{
\item{object}{a fitted model object}

\item{parsed}{an interpreted \code{dlm} formula object returned by
\code{\link{interpret.dlm}}}

\item{call}{an optional matched function call}
}
\value{
an S4 object of class \code{\link{dlMod}}
}
\description{
Convert an appropriately fit model object into an object of class
\code{\link{dlMod}}
}
