% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDlm}
\alias{plotDlm}
\alias{plot.dlMod}
\title{Plot smoothed lag terms}
\usage{
plot(x, geom = c("pointrange", "line"), level = 0.95, scaled = TRUE, ...)
}
\value{
a \code{ggplot2} graphic object
}
\description{
Plot estimated lag-spline effects
}
