% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-LagBasis.R
\name{predict.LagBasis}
\alias{predict.LagBasis}
\title{Predict new values for fitted lag basis}
\usage{
\method{predict}{LagBasis}(object, x, ...)
}
\arguments{
\item{object}{A \code{\link{LagBasis}} object}
}
\description{
Will update w/ description. Not yet implemented
}
