% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/methods-dlMod.R
\docType{methods}
\name{scaleMat}
\alias{scaleMat}
\alias{scaleMat}
\title{Extract distributed lag scale matrix}
\usage{
scaleMat(object, ...)

\S4method{scaleMat}{dlMod}(object, ...)
}
\arguments{
\item{object}{a fitted model object}
}
\value{
A square numeric matrix represented as a
\code{Matrix::\link[Matrix]{sparseMatrix}} object
}
\description{
Return lag coefficient scale matrix, \eqn{S}. \eqn{S} should be invertable.
Typically users should only interact with \code{scaleMat} indirectly
through the \code{\link{estimands}} functions (with the argument
\code{scaled = TRUE}).
}
\details{
If \eqn{\theta} is the vector of unscaled regression coefficients,
then corresponding distributed lag terms in the vector
\eqn{S \theta}{S * \theta} can be interpreted as average regression
coefficients up to their associated radius. Non-DL coefficients are
unchanged by this transformation.
}
