% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{se.fixef}
\alias{se.fixef}
\title{Extract Standard Errors of Fixed Effects}
\usage{
se.fixef(object)
}
\value{
A numeric vector
}
\description{
Returns a vector of fixed effects standard errors from a fitted
mixed effects model.
}
\details{
\code{fixef} is S4 generic.
}
