% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{se.ranef}
\alias{se.ranef}
\title{Extract Standard Errors of Random Effects}
\usage{
se.ranef(object)
}
\value{
A numeric vector
}
\description{
Returns a vector of random effects standard errors from a fitted
mixed effects model.
}
\details{
\code{fixef} is S4 generic.
}
