% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm-functions.R
\name{secumsum}
\alias{secumsum}
\title{Standard Error Cumulative Sum}
\usage{
secumsum(vcov)
}
\arguments{
\item{vcov}{A variance-covariance matrix}
}
\value{
A vector of SEs
}
\description{
Gives the cumulative sum of a variance-covariance matrix.
}
