% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/methods-dlMod.R
\docType{methods}
\name{Sigma}
\alias{Sigma}
\alias{Sigma,dlMod-method}
\title{Extract model coefficients variance matrix}
\usage{
Sigma(object, ...)

\S4method{Sigma}{dlMod}(object, scaled = TRUE, ...)
}
\section{Methods (by class)}{
\itemize{
\item \code{dlMod}: S4 Method for \code{\link{dlMod}} objects
}}

