% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bases.R
\name{smoothing}
\alias{smoothing}
\alias{cr}
\alias{sm}
\title{Basis smoothing}
\usage{
cr(x, Z, ...)

sm(x, Z, ..., .fun = NULL)
}
\arguments{
\item{x}{a vector of values to construct the basis from. Missing values
are not allowed}

\item{Z}{a covariate matrix (or object that can be coerced to a \code{matrix})
to apply the linear basis transformation to.
\code{length(x)} should be the same as \code{ncol(Z)}}

\item{...}{arguments to be passed to \code{\link{basis}}}
}
\value{
An S4 object of class \code{\link{SmoothLag}}.
}
\description{
Constructs a set of basis vectors for distances between distributed
lag points, and apply as a linear transformation of a
concentration matrix.
}
\details{
These functions are little more than convenient
  wrappers to the function \code{\link{basis}} and the
  \code{\link{SmoothLag}} class constructor. They are intended to
  simplify the task of specifying lag terms in a model \code{formula}.
  The functions compute a set of basis vectors for
  parameter \code{x} and applies this basis as a linear transformation
  of the covariate/concentration matrix parameter, \code{Z}. For example,
  if \code{cr} is used and \code{Z} is the identity matrix,
  the model fit will simply be the natural cubic spline of \code{x}.

  Note that other basis extensions should always return an object
  that inherits from \code{\link{SmoothLag}}
}
\section{Functions}{
\itemize{
\item \code{cr}: natural cubic radial basis spline

\item \code{sm}: user-defined smoothing
}}

\examples{
## load simulated data set and extract concentration matrix
data (simdata)
Conc <- simdata[, -(1:3)]  # First columns are Y, Age, and Gender

## radial lag (distance) each concentration was measured at
x <- seq(0.1, 10, length.out = ncol(Conc))
crb <- cr(x, Conc)

}
\references{
Rupert D, Wand MP, & Carroll RJ (2003) Semiparametric
Regression. New York: Cambridge University Press.
}
\seealso{
\code{\link{basis}}, \code{\link{SmoothLag}}
}
