% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm-functions.R
\name{standard_twfe_for_comparison}
\alias{standard_twfe_for_comparison}
\title{Standard Two-Way Fixed Effects Model for Comparison}
\usage{
standard_twfe_for_comparison(
  data,
  from_rt,
  to_rt,
  outcome,
  time,
  unit,
  time_to_treatment,
  treat,
  covariates = NULL,
  ref_period = -1
)
}
\arguments{
\item{data}{A data frame containing the main dataset.}

\item{from_rt}{The starting lag period.}

\item{to_rt}{The ending lag period.}

\item{outcome}{The outcome variable.}

\item{time}{The time variable.}

\item{unit}{The unit identifier.}

\item{time_to_treatment}{The variable representing time to treatment.}

\item{treat}{The treatment indicator variable.}

\item{covariates}{Optional covariates for the model.}

\item{ref_period}{The reference period for summing coefficients (default is -1).}
}
\value{
A list containing model results, coefficients, and plots.
}
\description{
This produces the standard two-way fixed effects model for comparison to the distributed lags model
}
