% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/methods-dlMod.R
\docType{methods}
\name{vcoef}
\alias{vcoef}
\alias{vcoef0,dlMod-method}
\alias{vcoef,dlMod-method}
\title{Vectorized coefficients}
\usage{
vcoef(object, ...)

\S4method{vcoef0}{dlMod}(object, scaled = TRUE, ...)

\S4method{vcoef}{dlMod}(object, scaled = TRUE, ...)
}
\value{
A numeric vector: with variable names in the case of \code{vcoef}
}
\description{
Extract fixed and random effects coefficient vector,
\eqn{(\beta, b)'} from a fitted \code{\link{dlMod}} object
}
\section{Methods (by class)}{
\itemize{
\item \code{dlMod}: S4 Method for \code{"\link[=dlMod]{dlMod}"} Objects

\item \code{dlMod}: S4 Method for \code{"\link[=dlMod]{dlMod}"} Objects
}}

