% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bases.R
\name{basis}
\alias{basis}
\title{Basis vector sets}
\usage{
basis(x, center = TRUE, scale = FALSE, ...,
      .fun = function(x, ...) abs(outer(x, x, "-"))^3)
}
\arguments{
\item{x}{a set of points to measure distances between; the resultant
distance matrix will be decomposed into a set of basis vectors}

\item{center}{if \code{TRUE} (the default), parameter \code{x} will be mean
centered prior to computing distances. Otherwise, if given a
\code{numeric} value, \code{x} will be centered at \code{center}}

\item{scale}{if \code{TRUE} (default = \code{FALSE}), parameter \code{x}
will be scaled by \code{sd(x)}. Otherwise, if given a
\code{numeric} value, \code{x} will be scaled by \code{scale}}

\item{...}{other parameters passed to \code{.fun}}

\item{.fun}{a function to compute the distances between the values in \code{x}.
The default is to compute pairwise cubed absolute distances}
}
\value{
An object of class \code{\link{LagBasis}}
}
\description{
Construct a set of basis vectors based on the distances between
input points

Users should not typically interact with \code{basis} directly.
Typical usage relies on calling basis application functions, like
\code{\link{cr}}, directly and/or in \code{\link{dlm}} model
formulae
}
\details{
Alternative distance functions, \code{.fun}, may be specified, and
error checking on the user's choice of \code{.fun} is deliberately
missing. Proper candidates for \code{.fun} should return a
(\code{length(x)} by \code{length(x)}) matrix, however, and values
are typically non-negative.

The distance matrix decomposition follows Rupert, Wand, and Carroll
(2003). In particular, once \code{x} and \code{.fun} are chosen and
we define distance matrix \code{C_1 = .fun(x, ...)}, and we let

\deqn{C_0 = [1^{(n \times 1)}, x]}{C[0] = [1, x]}
\deqn{C_1 = Q R}{C[1] = Q * R}
\deqn{M_1 = Q_{-(1:2)}}{M[1] = Q[-(1:2)]}
\deqn{K_1 = C_1 M_1 (M_1^T C_1 M_1)^{-\frac{1}{2}}}{K[1] = C[1] * M[1] * (M[1]' * C[1] * M[1])^-0.5}

where \eqn{A_{-j}}{A[-j]} denotes a matrix \eqn{A} with column(s) \eqn{j}
removed. Then we are interested in estimating effects scaled by the
matrix \eqn{\Omega = [C_0, K_1]}{\Omega = [C[0], K[1]]}
}
\references{
Rupert D, Wand MP, & Carroll RJ (2003) Semiparametric
Regression. New York: Cambridge University Press.
}
\seealso{
\code{\link{cr}}, \code{\link{dlm}}
}
