% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bases.R
\name{basis}
\alias{basis}
\title{Basis vector sets}
\usage{
basis(x, center = TRUE, scale = FALSE, .fun = NULL, ...)
}
\arguments{
\item{x}{a vector or radii; the resultant
distance matrix will be decomposed into a set of basis vectors}

\item{center}{if \code{TRUE} (the default), parameter \code{x} will be mean
centered prior to computing distances. Otherwise, if given a
\code{numeric} value, \code{x} will be centered at \code{center}}

\item{scale}{if \code{TRUE} (default = \code{FALSE}), parameter \code{x}
will be scaled by \code{sd(x)}. Otherwise, if given a
\code{numeric} value, \code{x} will be scaled by \code{scale}}

\item{.fun}{a function to compute distances between the values in \code{x}.
The default is to compute pairwise cubed absolute distances. See
Details}

\item{...}{other parameters passed to \code{.fun}}
}
\value{
An object of class \code{\link{LagBasis}}
}
\description{
Construct a set of basis vectors based on the distances between
input points.
Typical usage relies on calling basis application functions, like
\code{\link{cr}} (e.g. in \code{\link{dlm}} model
formulas); users should not often have to interact with \code{basis}
directly.
}
\details{
Alternative distance functions, \code{.fun}, may be specified, and
error checking on the user's choice of \code{.fun} is deliberately
missing. Proper candidates for \code{.fun} should return a
(\code{length(x)} by \code{length(x)}) matrix, however, and values
are typically non-negative.

In addition, new distance function definitions should follow the idiom:

\code{function(x, y, ...)}

\code{  if (missing(y))  y <- x}

\code{  ...}

The default value of \code{.fun} computes cubic radial distance,
which amounts to \code{abs(outer(x, y, "-"))^3}; distance matrix
decomposition follows Rupert, Wand, and Carroll
(2003).
}
\section{Decomposition}{

Once a distance function (\eqn{\delta(\cdot, \cdot)}{\delta()}) and
radii (\eqn{r}) are chosen, define distance matrix,
\eqn{C_{1, ij} = \delta(r_i, r_j)}{C_1[i, j] = \delta(r_i, r_j)},
and let,

\deqn{C_0 = [1^{(n \times 1)}, r]}{C_0 = [1, r]}
\deqn{C_1 = Q R}{C_1 = Q * R}
\deqn{M_1 = Q_{-(1:2)}}{M_1 = Q[-(1:2)]}
\deqn{K_1 = C_1 M_1 (M_1^T C_1 M_1)^{-\frac{1}{2}}}{K_1 = C_1 * M_1 * (M_1' * C_1 * M_1)^-0.5}

where \eqn{A_{-j}}{A[-j]} denotes a matrix \eqn{A} with column(s) \eqn{j}
removed. Then distributed lag effects of interest are scaled by the
matrix \eqn{\Omega = [C_0, K_1]}{\Omega = [C_0, K_1]}
}

\references{
Rupert D, Wand MP, & Carroll RJ (2003) Semiparametric
Regression. New York: Cambridge University Press.
}
\seealso{
\code{\link{cr}}, \code{\link{dlm}}
}
