% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\docType{methods}
\name{cholfVar}
\alias{cholfVar}
\alias{cholfVar,merMod-method}
\title{Extract Cholesky factor of inverse Information matrix}
\usage{
cholfVar(object, ...)

\S4method{cholfVar}{merMod}(object, ...)
}
\section{Methods (by class)}{
\itemize{
\item \code{merMod}: S4 Method for \code{lme4::\link[lme4]{merMod}} objects
}}

