% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm-functions.R
\name{distributed_lags_models}
\alias{distributed_lags_models}
\title{Distributed Lags Models}
\usage{
distributed_lags_models(
  data,
  exposure_data,
  from_rt,
  to_rt,
  outcomes,
  exposure,
  unit,
  time,
  covariates = NULL,
  addl_fes = NULL,
  ref_period = -1,
  weights = NULL,
  dd = F,
  n = 2,
  dict = NULL,
  remove_unit_FE = FALSE,
  addl_arguments = c(),
  model_type = "feols"
)
}
\arguments{
\item{data}{A data frame containing the unit, time, outcome, covariates, and any additional fixed effects}

\item{exposure_data}{A data frame containing the unit, time, and exposure variables}

\item{from_rt}{The starting lag period.}

\item{to_rt}{The ending lag period.}

\item{exposure}{The exposure variable.}

\item{unit}{The unit identifier.}

\item{time}{The time variable.}

\item{covariates}{Vector of covariates for the model.}

\item{addl_fes}{Vector of additional fixed effects for the model.}

\item{ref_period}{Reference period (default -1)}

\item{weights}{Weights to be included in the regression}

\item{dd}{Whether to include the DD estimate in the plot}

\item{n}{Number of digits to round to (for the DD estimates)}

\item{dict}{A dictionary of variable names}

\item{remove_unit_FE}{Whether to remove the unit fixed effects}

\item{addl_arguments}{Additional arguments to be included in the model, as strings}

\item{outcome}{The outcome variable.}
}
\value{
A list containing model results, coefficients, and plots.
}
\description{
This is the distributed lags model / continuous event study. It allows for multiple outcomes.
}
