% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm.R, R/zzz.R
\docType{package}
\name{dlm}
\alias{dlm}
\alias{dlm}
\alias{dlm-package}
\title{Distributed lag models}
\usage{
dlm(formula, data, subset, na.action, weights, offset,
    method = c("REML", "MLE"), family = gaussian(),
    control = list(), ...)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}}:
a symbolic description of the model to be fitted. See Details.}

\item{data}{an optional data frame, list, or environment containing the
variables of the fitted model.}

\item{subset}{optional vector specifying a subset of observations to be used in the
fitting process.}

\item{na.action}{optional function that indicates what should happen when the data contains
\code{NA}'s.}

\item{weights}{optional vector of weights to be used in the fitting process. Should be
\code{NULL} or a numeric vector}

\item{offset}{a known offset term to include in the model}

\item{method}{algorithm used to fit the DLM. Partial matching and capitalization allowed}

\item{family}{a \code{\link[stats]{glm}} family}

\item{control}{a list of simulation control parameters. See Details}

\item{...}{Additional parameters passed to \code{\link[lme4]{lFormula}}}
}
\value{
An S4 object that inherits from \code{\link{dlMod}} and
\code{\link[lme4]{merMod}} containing the results of the fitted model.
Many standard model summary methods are available for these object
types
}
\description{
Fit distributed lag models using \pkg{lme4} to penalize smooth terms.
Other random effects terms and generalized mixed models supported.

DESCRIPTION OF PACKAGE HERE!
}
\examples{
data (simdata)

## Setup distance count matrix and corresponding lag distances
X <- as.matrix(simdata[, -(1:3)])
lag <- seq(0.1, 10, length.out = ncol(X))

fit <- dlm(Y ~ Age + Gender + cr(lag, X), data = simdata)
summary (fit)

}
\references{
Baek J, Sanchez BN, Berrocal BJ, & Sanchez-Vaznaugh EV
(2016) Epidemiology 27(1):116-24.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/26414942}{PubMed})

Baek J, Hirsch JA, Moore K, Tabb LP, et al. (2017)
Epidemiology 28(3):403-11.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/28145983}{PubMed})

Bates D, Maechler M, Bolker BM, & Walker SC (2015)
Fitting linear mixed-effects models using lme4. J
Stat Softw 67(1).
(\href{https://www.jstatsoft.org/article/view/v067i01/0}{jstatsoft.org})
}
\seealso{
\code{\link[lme4]{lmer}}, \code{\link{cr}},
  \code{\link{dlMod}}
}
