% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm.R
\name{dlm}
\alias{dlm}
\title{Distributed lag models}
\usage{
dlm(formula, data, subset, na.action, weights, offset,
    method = c("REML", "MLE"), family = gaussian(),
    control = list(), ...)
}
\arguments{
\item{formula}{an object of class \code{stats::\link[stats]{formula}}:
a symbolic description of the model to be fitted. See Details}

\item{data}{an optional data frame, list, or environment containing the
variables of the model to be fitted}

\item{subset}{optional vector specifying a subset of observations to be used in the
fitting process}

\item{na.action}{optional function that indicates what should happen when the data contains
\code{NA}'s. The default is set by the \code{na.action} setting of
\code{base::\link[base]{options}}, usually
\code{stats::\link[stats]{na.omit}}}

\item{weights}{optional vector of weights to be used in the fitting process. Should be
\code{NULL} or a numeric vector}

\item{offset}{a known offset term to include in the model, e.g. for \code{poisson()}
family models}

\item{method}{algorithm used to fit the DLM. Partial matching and capitalization allowed.
The default is \code{"REML"} for linear/\code{gaussian(link = "identity")}
family models, and \code{"MLE"} otherwise}

\item{family}{a description of the error distribution and link function to be used
in the model. The default is \code{gaussian(link = "identity")}.
See \code{stats::\link[stats]{family}} for possible family
functions and details}

\item{control}{either a \code{list} object with arguments to be passed to the
\code{lme4::\link[lme4]{lmerControl}} sequence, or the output of
\code{[g]lmerConrol} directly}

\item{...}{Additional parameters passed to \code{lme4::\link[lme4]{lFormula}}}
}
\value{
An S4 object that inherits from \code{\link{dlMod}} and
\code{lme4::\link[lme4]{merMod}} containing the results of the fitted model.
Many standard model summary methods are available for these object
types
}
\description{
Fit distributed lag models using \pkg{lme4} to penalize smooth terms.
Other random effects terms and generalized mixed models supported.
}
\details{
Models are specified using typical \pkg{lme4} \code{formula} syntax with
at least one set of
lag terms returned by a given smoothing function (e.g. see \code{\link{cr}}).
The smoothing function can be any that returns a \code{\link{SmoothLag}} basis
object. See Examples
for a basic call to \code{dlm} using the formula interface, and a cubic
radial lag basis specified via \code{cr}, and the \pkg{dlmBE}
\code{\link[=dlmBE-package]{package documentation}} for a discussion
of the types of models \code{dlm} is designed to handle.
}
\examples{
data (simdata)

## Setup distance count matrix and corresponding lag distances
X <- as.matrix(simdata[, -(1:3)])
lag <- seq(0.1, 10, length.out = ncol(X))

fit <- dlm(Y ~ Age + Gender + cr(lag, X), data = simdata)
summary (fit)

}
\references{
Baek J, Sanchez BN, Berrocal VJ, & Sanchez-Vaznaugh EV
(2016) Epidemiology 27(1):116-24.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/26414942}{PubMed})

Baek J, Hirsch JA, Moore K, Tabb LP, et al. (2017)
Epidemiology 28(3):403-11.
(\href{https://www.ncbi.nlm.nih.gov/pubmed/28145983}{PubMed})

Bates D, Maechler M, Bolker BM, & Walker SC (2015)
Fitting linear mixed-effects models using lme4. J
Stat Softw 67(1).
(\href{https://www.jstatsoft.org/article/view/v067i01/0}{jstatsoft.org})
}
\seealso{
\code{lme4::\link[lme4]{lmer}}, \code{\link{cr}},
  \code{\link{dlMod}}
}
