% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{dlMod}
\alias{dlMod}
\title{Distributed lag model objects}
\description{
A fitted distributed lag model object. Inherits from \pkg{lme4}'s
\code{\link[lme4]{merMod}} so that most methods defined for this
parent class should work seamlessly within \pkg{dlmBE} analysis
}
\section{Slots}{

\describe{
\item{\code{resp}}{An \code{lme4::\link[lme4]{lmResp}} object to store a
(mixed) model response variable}

\item{\code{bases}}{A list of \code{\link{LagBasis}} objects corresponding
to the unique set of lag bases used to fit the model}

\item{\code{index}}{A (named) integer vector providing the index of the
basis set in \code{bases} corresponding to each lag term in the
model}
}}

