% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{LagBasis}
\alias{LagBasis}
\title{Create and Manipulate Lag Basis Functions}
\description{
S4 class object to store and query components of
lag basis functions. User interface for creating this class can be found in
\code{\link{basis}}.
}
\details{
See \code{\link{basis}} for details of the decomposition
}
\section{Slots}{

\describe{
\item{\code{x}}{original lag data}

\item{\code{x.center}}{store the value the lag data was centered to. In theory
this is useful for the non-yet-implemented \code{predict} method}

\item{\code{x.scale}}{store the value the lag data was scaled by. Again, should be
useful primarily for the \code{predict} method in the future}

\item{\code{C0}}{C_0 part of basis matrix}

\item{\code{K1}}{K_1 part of basis matrix}
}}

\references{
Rupert D, Wand MP, & Carroll RJ (2003) Semiparametric
Regression. New York: Cambridge University Press.
}
