% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/methods-LagBasis.R
\docType{methods}
\name{omega}
\alias{omega}
\alias{omega,LagBasis-method}
\title{Extract Lag Basis Matrix}
\usage{
omega(object, ...)

\S4method{omega}{LagBasis}(object, ...)
}
\value{
A square numeric matrix
}
\description{
Extract lag basis matrix, \eqn{\Omega}
}
\details{
\code{omega} is S4 generic.
}
\section{Methods (by class)}{
\itemize{
\item \code{LagBasis}: Method for "\code{LagBasis}" objects
}}

