% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{predict.LagBasis}
\alias{predict.LagBasis}
\alias{predict.LagBasis}
\title{Predict New Values for Fitted Lag Basis}
\usage{
\method{predict}{LagBasis}(object, ...)

\method{predict}{LagBasis}(object, ...)
}
\arguments{
\item{object}{A \code{\link{LagBasis}} object}

\item{object}{A \code{\link{LagBasis}} object}
}
\description{
Not yet implemented

Not yet implemented
}
