% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm-functions.R
\name{serevcumsum}
\alias{serevcumsum}
\title{Reverse Standard Error Cumulative Sum}
\usage{
serevcumsum(cov)
}
\arguments{
\item{cov}{A variance-covariance matrix}
}
\value{
A vector of SEs
}
\description{
Gives the cumulative sum of a variance-covariance matrix.
}
