% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SmoothLag}
\alias{SmoothLag}
\title{Lag Matrix With Applied Smoothing}
\description{
An S4 object for representing lag covariates and
storing details about the basis smoothing. Intended for use within the
\code{\link{dlm}} modeling framework to assist extraction of basis
components treated as "fixed" and "random" effects. Inherits from
\code{matrix}.
}
\section{Slots}{

\describe{
\item{\code{basis}}{A \code{\link{LagBasis}} smoothing object containing details about the
lag and the smoothing parameters used}

\item{\code{.Data}}{Contains the "fixed effects" components of the smoothed lag function.
This scheme is intended to work conveniently with
\code{stats::\link[stats]{model.matrix}}}

\item{\code{random}}{Contains the random effects or penalized components of the smoothed
lag function}

\item{\code{terms}}{Character vector containing the function name and deparsed arguments
from whatever smoothing function generated the \code{SmoothLag} object}
}}

